+++ Source/Sitecore.FakeDb.Tests/ProviderBehaviorSwitcherTest.cs
@@ -1,0 +1,16 @@
+﻿namespace Sitecore.FakeDb.Tests
+{
+  using FluentAssertions;
+  using NSubstitute;
+  using Sitecore.FakeDb.Links;
+  using Sitecore.Links;
+  using Xunit;
+
+  public class ProviderBehaviorSwithcerTest
+  {
+    [Fact]
+    public void ShouldTemporarilyOverrideProviderBehavior()
+    {
+    }
+  }
+}