+++ Source/Sitecore.FakeDb/IBehavioral.cs
@@ -1,0 +1,7 @@
+﻿namespace Sitecore.FakeDb
+{
+  public interface IBehavioral<TProvider>
+  {
+    TProvider Behavior { get; set; }
+  }
+}