+++ Source/Sitecore.FakeDb/Links/LinkDatabaseSwitcher.cs
@@ -1,27 +1,12 @@
 ﻿namespace Sitecore.FakeDb.Links
 {
-  using System;
-  using Sitecore.Diagnostics;
   using Sitecore.Links;
 
-  public class LinkDatabaseSwitcher : IDisposable
+  public class LinkDatabaseSwitcher : ProviderBehaviorSwitcher<LinkDatabase>
   {
-    private readonly FakeLinkDatabase _linkDatabase;
-
     public LinkDatabaseSwitcher(LinkDatabase behavior)
+      : base((FakeLinkDatabase)Globals.LinkDatabase, behavior)
     {
-      Assert.ArgumentNotNull(behavior, "behavior");
-
-      var linkDb = Globals.LinkDatabase as FakeLinkDatabase;
-      Assert.IsNotNull(linkDb, "linkDb");
-
-      this._linkDatabase = linkDb;
-      this._linkDatabase.Behavior = behavior;
-    }
-
-    public void Dispose()
-    {
-      this._linkDatabase.Behavior = null;
     }
   }
 }