+++ Source/Sitecore.FakeDb/Resources/Media/MediaProviderSwitcher.cs
@@ -1,0 +1,12 @@
+﻿namespace Sitecore.FakeDb.Resources.Media
+{
+  using Sitecore.Resources.Media;
+
+  public class MediaProviderSwitcher : ProviderBehaviorSwitcher<MediaProvider>
+  {
+    public MediaProviderSwitcher(MediaProvider behavior)
+      : base((IBehavioral<MediaProvider>)MediaManager.Provider, behavior)
+    {
+    }
+  }
+}